pyproject.toml
[project]
name = "pythonid-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-telegram-bot[job-queue]>=22.5",
    "sqlmodel>=0.0.28",
]

[project.scripts]
pythonid-bot = "bot.main:main"

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bot"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.coverage.run]
source = ["src/bot"]
omit = ["*/tests/*", "src/bot/main.py", "*/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "if __name__ == .__main__.:",
]
